ht: add new check RenderingTime
How long it really takes to load all page assets and draw the page might
be something worth to check. Add a check because it is so easy even if
it will be hard to enforce hard limits which such a check as this rendering
time probably show a very broad histogram.

Unrelated but slipped into this commit: Produce better documentation.