cmd/ht: terminate output on stdout with overall status
The report printed starts with the status of the suite (pass, ... bogus) and it
ends with a summary line how many test finished in each status. Even for moderately
longsuites the start of the printed report will scrol out of the terminal and
inspecting the summary line requires higher brain functions.
Report the overall outcome (PASS, FAIL, ERROR, BOGUS) on the last line as a
human readable exit code.