cmd/ht: add new subcommand 'doc'
The new subcommand just invokes 'go doc github.com/vdobler/ht/ht <type>'
and slightly cleans up its output. This allows to look up the details
of a check, the Test, Request or a variable extractor as long as the
ht command does not import other packages for their checks/extractors.
If this happens it might be possible to scna the check and extractor
registries to llok up the package name. Not a concern now.