ht: don't provoke header failures in resilience check
Some things simply cannot be sent as header values at least not through
net/http. It is nonsensical to provoke errors which are not dependent
on the server.
Add a new category of modifications 'space' which contains different
whitespace strings and 0 bytes and do not use this category for
header modifications.