recorder: full rewriting of everything
Some web aplications are just ugly to spy on which is okay
as these problems stem vom CSRF protections and setting
cookies via JavaScript and that like so we will have to deal
with it.

This change allows to completely rewrite all headers and the
body if both, the requests and the responses.
For responses it is necessaryy, for requests it is experimental.
The replacement is a plain textual replacement.

This works pretty well if you can listen on port 80: In this
case you local address doesn't need a port number and this
replacement works totaly fine. Just add something to your host
file and record as root :-)