ht: forward Basic Auth header on redirects
Following redirects on a host which requires Basic Auth was broken as the
needed Authentication header was sent only in the first request but not
the onews from following the redirection. This makes the whole
FllowRedirects stuff useless and undermines the whole idea of ht.

Forward to the same host if set.