ht: add short reports
The default text reports got longer and longer; which is nice as you get
a lot of information what happened and what went wrong but these
reports never fit on a single screen and one has to scroll all the
time.
Thus condense the reports, show name and request only and what
whent wrong (if anything went wrong).