ht: turn off chunked POST bodies per default
When in doubt: Do as curl does.
This CL tries to fix a strange bug found while testing a PHP application:
The same request with curl worked while with ht it failed as no request
body was visible inside the PHP application for POST requests with a body.
The only real difference was: curl didn't do chunked transfer encoding
but properly announced the content length of the following body and sent
it as a whole.
Do this with ht as default. To allow chunked request body add a new
field Chunked to Test.Request.

This is a bit experimental.