ht: improve checking logfiles on remote machines
Persist only the client config (instead of a full connection) during Prepare.
It is okay if a remote file is inexistent during Prepare.
Handle missing remote files during execute.
Improve documentation and make a proper test (which is skipped if no
user name and credentials are given on the cmdline.)