ht: add new check CustomJS
Sometimes the set of provided checks just is not enough or not
flexible enough. If theis happens you cannot use the ht command to
validate the response.
This CL adds a new type of Check which executes custom JavaScript
and has access to the Test including Request and Response. This makes
it possible to program your own check by coding the check in JavaScript.

JavaScript is interpreted by otto (github.com/robertkrimen/otto).