cmd/ht: ignore teardown tests in global status
A suites status has always been the maximum status of the setup
and the main tests, ignoring the status of the teardown tests.
This seems pretty sensible. Strangly enough the exec command
included the teardown tests when reporting the overall status of
executing one or several suites. This seems to be a bad idea as
teardwon errors might pollute the overall outcome.

Exec still counts the outcome of the teardown requests (and will
happily display 'Errored: 2') but will ignore teardown request
while computing the overall PASS/FAIL/ERROR message and the
exit code.