cmd/ht: allow saving variable state after execution
This CL allows to save the values of all the variables after executing a suite
in JSON form. Actually the variables are written to the file 'variables.json' in
the result folder and optionally all variables consolidated over all suites
to a file specified via the -dump cmdline option.

This is useful for debugging purpuse but also to resume a suite or split one
long suite into to half (and do something else between the two halfs).
What is still missing is saving and restoring cookies.