cmd/ht: allow dumping and loading ofand loading of variables and cookies
Splitting a suite into a first half, doing some bash scripting and
then executing the second half of the suite was basically impossible
to do. To allow this one has to carry over the state of a suite
which are the current set of variables and the cookies. Allow both to
be written to disk after execution of a suite. Loading variables from
a file was already possible with -Dfile, add this functionality for
cookies. (Technicall this could be done as a Test's Request allows
to set cookies, but this is awkward).
Add some tests for cmd/ht invocation.