ht: save response body with proper filename extension
Today I learend that there are people who open files by clicking on the filename
(instead of the normal less, od | less or via emacs). Strange but true.
To faciliate the live of these people: Save the response body while generating
the report in a file with the proper extension.
The extension is taken from the Content-Type header or it is .txt.