populate: don't populate maps from slices
Populating a map from a slice seems like a nice idea: Index becomes key,
Value maps to Value. But in combination with hjson this produces funny
results which are hard to debug:

  Foo: [
    key: value
  ]

is a hjson array with one element "key: value", populate that to a
map Foo you end up with a map with one entry:

  "0": "key: value"

which probably is not wht you wanted.