ht: Consistent naming of elements in JSON
The JSON check and the JSONExtractor both work on exploded JSONs. While the first
used 'Element' to select which atomic field should be checked, the second used
'Path' to select the field to extract. This is stupid and painful to work with:
You often check for the existence (and maybe the welformednesss) of an field
and extract it afterwards.
Rename Path to Element in JSONExtractor.

This is an incompatible change and will break tests.