vendor: do not convert ints to floats in gojsonexplode
Gojsonexplode roundtripps form JSON to Go and back. Now there are no int in JSON and the
JSON-->Go produces a float when unmarshalling a number. If this number is large, than
the Go-->JSON remarshalling will format it as a floating point numer, e.g. 1.23456e6 which
is ugly and breaks this roundtripping.

Quickfixhack to keep int64 if not a 'real' float in the JSON.