suite: remove (external) JavaScript dependecy from HTML report
The old HTML report used jQuery to display/hide stuff in the report.
This made using the reports hard to use if the external jQuery could not be
loaded, e.g. because WiFi is unavailable.

Make the show/hide stuff CSS based only. While at it clean up the
generated HTML a bit to make the W3C validator a bit happier.