ht: allow checking of JSON embeded in JSON
There seems to be a trend in transmitting JSON structures serialized as a string and
wrapped in an outer (real) JSON. The original JOSN check just sees a string and checking
inside this string is hard and eror prone. This CL unwraps such embeded JSON and
applies a full JSON test.
This allows to test all these fancy modern variants of JSON. This is progress!
(Not sure I actually do understand why everybody likes this russion-doll-JSONs
but what do I know about modern stuff...)