ht: fix generating screenshots of basic-auth protected pages
PhantomJS is used to render the actual HTML page and PhantomJS has to request the
linked CSS and Javascript files to layout and render the page. If these assets
are protected by Basic Auth than PhantomJS has problems loading them and the resulting
screenshot will look strange (or the page cannot be rendered at all because of
JavaScript failures).
So we must tell PhantomJS to authenticate its requests to resources. Unfortunately
the simple solution of using page.setting.{userName,password} seems to be broken
(see http://darrendev.blogspot.ch/2013/04/phantomjs-post-auth-and-timeouts.html).
Adding cutom header might be dangerous as they might be sent to all linked resources,
even resources on different hosts but there seems to be no simple solution.