ht: wait for animations before making screenshots
Taking a screenshot directly after page load might not capture stuff rendered async
into the page. It was always possible to execute a script which could wait, but this
is a major pain (given that there is no non-busy, synchronious sleep in JavaScript).

Incorporate such waiting for elements becoming visible/invisible into the standard code.

While at it make the JavaScript code pretty generic so that the other two PhantomJS
based checks (RenderingTime and RenderedHTML) might use this code too.