ht: consolidate all PhantomJS based tests
The RenderedHTML and RenderingTime checks did not take a screen geometry which is
a bug as the generated HTML as well as the time to render it might depend on
screensize and viewport.
Alos for both cecks it might be interesting to wait until some element in the
page becomes (in)visible before evaluating the HTML or declaring the page rendered.

Extract this functionality from Screenshot into a new struct Browser end embed it
in all three checks. Make all three checks use basically the same PhantomJS driving
script.