ht: add 'bash://' pseud-request
Sometimes it would be handy to execute some shell script in the middle of a suite.
This changes adds the possibility to execute some shell script (passed as the Request.Body)
in a certain directory (passed as the URL.Path) and interprete the output as the response
and apply CHecks to it.

Very experimental. The idea is to add SQL-DB quieres too.

While at it sperate all pseudo request (not file:// and bash://) into their own source code file.