ht: improve and exapnd sql pseudoqueries
This change adds csv output and makes headers optional
(can be turned on in csv and plaintext via a "headers=present"
parameter to the media type). Plaintext seperator is now
setable via the media type parameter "fieldsep". The default
is Tab '\t' as this is hard to set via fieldsep.

As we need the Accept HTTP header anyway it makes sense to
pass in the data source name via the header too. Also spell
it out: Data-Source-Name is much more clear than DSN.

Tests and documentation have been improved.