cmd/ht: allow to load all suites with <dir>/... syntax
Ths CL adds no real new functionality, just some convenience, especially on
Windows where normal command line tooling might be unavailable or clumsy.

Any argument to the list and exec subcommand ending in "/..." are expanded
to all *.suite files below the basename of the argument. A bit like the go
tool handles three dots.

With this
    ht exec foo/...
Is equivalent to
    ht exec $(find foo -type f -name \*.suite | sort)
or (probably) in Java speak
    ht exec foo/**/*.suite

Fixes issue #8.