ht: allow export of a test request as a curl call
For debugging a single request in a suite it is simpler to execute this request
with curl from the command line (cmd/ht had a similiar subcommand once but ist
was basically unusable and was removed). Unfortunately it is hard to handcraft
a curl call from an executed test.
Add the possibility to automatically create a curl call to redo a request via
the command line.
Shell escaping is still suboptimal and might be buggy and unusable for binary
requst bodies (like e.g. a zip or png file).