all: experimental display of curl calls
Curl is a known standard and easier to use during debugging than ht due to the
faster turnaround times on the commandline. This change adds an experimental
feature: Constructing a curl command line call from a Test.Request to allow
easier debugging and fiddeling with the request on the command line.

This is experimental. The curl call is displayed in the generated HTML report
only.