suite: run teardown tests always
The teardown tests where skipped if any setup test failed. This produces
problems if e.g. the first 3 setup tests pass and produce stuff to clean up
but a failing fourth setup test skipps everything, including all teardown
tests.
Change this to the more intuitive: 'Teardown tests are run always!'.

While at it change the iteration logic: The executor is in a much better
and saner position to determine whether a test should be skipped than the
code doing the simple iteration.