suite: add verbosity field to Suite
The RawSuite already had this field, but this was used only to populate the
verbosity of the tests. As variable extraction and updating happens on a Suite
level there was no way to turn these dignostic messages off which is anoying
especially for load tests.
Add an exported Velocity field to Suite and output messages about updated variable
values only when the verbosity >=2.