suite: let each repetition in a load test have its own session
The old code was broken: All repetitions of a scenario in all threads used the same
cookie jar. The idea was that the Setup tests can prepopulate cookies but this
idea is braindamaged: Setup (and Teardown) in load tests are executed once only
before the different threads are started. Setup is good for populating variables
extracted from request but not for cookies which are typically tied to user
sessions.
Make every repetition of a scenario start afresh with an empty cookie jar.
Improve documentation while at it.