suite: rework load testing to be less racy and allow more load
Up to now each scenario used an unbuffered channel to sen test to the
request executor which means that one one thread working on a scenario
can run and prepare tests for execution (as the other threads are blocked
on sending to the unbuffered channel). This might limit the rate at which
we can generate request. Making this channel a bit buffered should make
it possible for several threads to prepare the next test which might
also lead to less thread misses.

Found a race with benders recording while debuging it, so fix it on
the fly.