ht: allow to disable Link test
The Which field of the Link check is mandatory. This makes it difficult to
disable thich check as one cannot simply use "" as an argument to which.
Allowing Which to be empty seems like a bad idea as this might result in
accidentialy disabled checks.
Allow (and ignore) the special value "-none-": Like this Which can be
non-empty but still omit all link checking.