all: safe loadtest data during test run
The old code did the whole loadtest run, accumulating everything in RAM
an dumping to disk only after finishing the test. This is bad because you cannot
do endurance tests like this (mitigated a bit by the fact that you could limit
collection to test which are bogus i.e. hopefuly not many), you cannot 'watch'
the results while the test is running and if a problem occurs (e.g. a panic) all
the work is lost.

So implement a streaming data writing solution. The output lacks some data
present in the full througput.csv as these columsn are calculated and the live
data is not sorted by starting time (but by ending time as this is when we
know the result). But it is better than nothing.