suite: allow abortion of load test on too high error rate
It is basically nonsense to keep a load/throughput test running once you killed the
system under test. The most sensible indicator that the system under test died is
the rate of errors. The rate of failures might be very interesting too: A proper system
might hide its internal dead behind HTTP status codes of 500. But for now looking at
the error rate only seems enough.
Errors are tracked over the last 50 terminating requests and the loadtest is aborted
of a certain threshold is exceeded.