asciistat: new package
This package is intended to replace hist. Drawing histograms via unicode
blocks seemd as a very clever idea. Unfortunately lots of fixed width
fonts do not have glyphs for this code point and those who do seem to be
not really fixed width: Adding a scale via 7-bit ascii results in
missalignments making the scale useless.

This package replaces a histogram with a boxplot-like graph. A standard
boxplot is drawn like this
    ---------[------M---]---------------
and larger percentiles (90, 95, 98 and 99) are indicated with |, ), }
and >.