ht: improve PhantomJS detection
The old code to detect whether PhantomJS is available (used to skip test
which rely on PhantomJS) had some drawbacks: It assumed any version of
PhantomJS would do (version 2 is a must), existence of a file would be
enough (it must execute too) and was potentialy racy. So replace it
with a nice sync.Once based solution which tries to actually execute
PhantomJS and read its version.