all: add offline load test analysation subcommand
It is nice if the final output of a loadtest is directly written to
disk in the final form but this is dangerous for soemthing like a
longrunning loadtest. During a loadtest we record the outcome as
a stream and write it to live.csv which contains all information
and can be used even if a throughput test crashes.

This CL adds a new subcommand 'stat' to cmd/ht to analyse the live.csv
written during a loadtest. It currently does nothing more than
converting it to the throughput.cv which is written once the load
test finishes.

The long term plan is to not write throughput.csv after a load test
and to enhance the stat subcommand to produce more or less data
from the live.csv (think of a filtered throughput.csv) and to implement
limits on the response times.