cmd/ht: make doc subcmd print list of checks and extractors
Up to now the help subcommand could print checks and extractors in some
strange, selfinvented pseudocode format. The real docstring could be printed
with 'go doc <type>'. So to list the available Extractors you hade to run
'go help' and to display the docstring 'go doc'. This is strange and the
incomplete pseudocode is not really helpful.

Make the 'go doc' subcommand responsible for listing available checks and
extractors and print just their names. This should be good enough to start further
investigation on the types (sill using 'go doc').

This is an incompatible change. As the help and doc subcommands are intended
for interactive use by humans this seems acceptable given that it is now
much pleasant to use.