ht: improve JSON documentation
The JSON check cannot be applied to non-leaf (internal) elements in the
JSON tree, only leaf elements have an elementary value to which a
Condition can be applied. This lead to ugly WTF moments when you try to
check that some JOSN contains a field foobar (which it does) but the
check fails with message 'element foobar not found'.

Be more specific in the documentation and the error message.