ht: improve JSON check
The old code uses gojsonexplode which has some drawbacks: It yields only leaf elements
and explodes the whole JSON. The first issue makes it impossible to test for mere
existense of an object/array in the JSON document, the second issue is just
unnecessary overhead (and garbage for the GC) as we are interested in one element
only.

The new code unwrapps the JSON step by step and decodes only the necessary parts.
Like this it also can look for non-leaf elements. Improve (at least I hope to do so)
the documentation and output better error messages.