ht: Remove *log.Logger and replace with a simple interface: Printf
Using a concrete type for logging has been proved by the community
as an anti-pattern ;-) So in this case we only need a single
function or more specific a behavior for logging. This behaviour
expresses: Printf(format string, a ...interface{}) and that's all
the logger needs to know.
Also removes the importing of the "log" package.