ht: add schema validation to JSON check
Sometimes it would be nice to check the overall structure of a JSON document,
maybe before diving into validating individual elements: If the structure is
broken the check for detailed values is probably useless.

Up to now there was no way to specify which structure the JSON under check
is expected to have. This CL introduces validation of the structure: The actual
JSON document is compared against a schema who's values just determine the
expected type (int, float, string, bool).