ht: use same code for JSON extraction as for JSON check
This really gets rid of the dependency to package gojsonexplode and allows to
extract non-leaf elements. It presumably is faster and produces less garbage
for large JSON files (which is important as each extraction just extracts one
element and it is common to extract several items from large JSON responses).