ht: implement typ checking for strings in Condition
Sometimes you do not care too much about the actual value received (e.g. in
a JSON document) but it must be a UUID variant 4 or an email address, or
an URL, or ... Testing this with regular expressions is hard, error prone
and unreadable.
This change includes (vendors) github.com/asaskevich/govalidator which
provides fifty plus checks, ranging from credit card number to hexcolor.
Checking happens on a logical OR basis to allow conditions of the form
"either an email address or a UUIDv4". (Logical AND is not provided as
this would be false almost always: No string is a UUID and an email.)

Checking for timestamps is also common, but here you need to know the
format. Instead of adding a dozen RFC1123Time, RFC3339Time, ...
Condition got a new filed Time which contains the expected time layout.

Both new conditions (Is and Time) remove double quotes from the string
before applying the condition: There are no URLs or UUIDs or Hexcolors
in double quotes. Such double quotes are typical during JSON checking
(because including them allows to check the type and differentiate
between 234 and "234") but these are useless for type and time checks.
This is a bit of an ugly hack and might indicate that returning the double
quotes in JSON was a bad idea.

While at it: Make the erros returned from Condition more homogenuous.