all: replace strange Reporting field in Test with general metadata
The Reporting section in a Test was only used to label/id different tests
so that nice HTML reports can be generated for a suite: It was part of the
(HTML) view of a Test. The throughput loadtest abused the SeqNo field to
transport information about the test through the load test. A second abuse.

This CL tries to replace this by a general mechanism of metadata: It is
possible to add (and read back) arbitrary meta data to a Test. This can be
used to attach view meta data (like sequence numbers, IDs, names) or
execution meta data (like scanario, thread and repetition number in a
load test).

As the abuse of SeqNo was pretty wild, it may be that the CL is not jet
bugfree or usable.