all: allow mocking of requests
This change introduces three things:
 - New package mock which allows a (hopefuly simple) definition of
   a mock response to a HTTP reuest.
 - New subcommand 'mock' to cmd/ht to execute such mocks and to debug
   them.
 - Integration of mocking into package suite, this is the most complex
   and described below.

The mocking works quite well but is far from finished. It should be
enough to use it, gain experience and see what is missing.

Integration into the execution of a suite works by naming the required
mocks when executing a test inside a suite. Before the test request is
made a server is started which will provide the mocked response.
The mocks themself may contain checks (which seems strange as the
response is mocked and not really subject to evaluation): The checks
inside a mock are used to check, to validate the incomming HTTP request.
The idea is to provide a real infrastructure to test a server:
The servers response is validated as well as any request to backend
systems.