all: introduce new package scope
It turned out that variable handling is common in packages suite and mock
which resulted in duplicate code or an import cycle. It turns out that
the variable handling stuff can be nicely seperated into a new package.

At the same time this allows proper variable replacement in a mock body.
To allow @vfile and @file syntax for the body some functionalities of
ht had to be exported.