all: refactor mock logging
The output of the mock subcommand to cmd/ht was just the test report (ASCII variant)
which is pretty uninformative. The log output during suite execution is much more
detailed, so refactor this type of output into a functions and use it in both
places.
The log report generated is unsuitable for binary formats or long text bodyieds, but
both seem uncommon to mock.