all: allow reset of global counter
The gloabl counter for the automatic variable COUNTER in a suite could not
be reset. This made testing a bit awkward bu the main problem happens in
cmd/ht: As suites are loaded and validated before beeing executed the
COUNTER did not start at 1 for the real execution. This is not a problem
per se but it seems wrong, especially as it starts with a value which
depends on how many suites/tests will be executed.

Add a mechanism to reset the counter to an arbitrary value.