all: use get rid of strings as IDs in load tests
The throughput tests in package suite relied on manipulating the test name to
transport information about scenario, test, thread and repetition through the
throughput test. This worked but is dead ugly and involes a lot of string
processing. With the new general metadata available in Test it is now possible
to do this in a typesafe way.
This change does not get rid of this string processing altogether, it is still
used during serialisation to disk.
While at it: Fix output file writing for cmd/ht stat, typos and other stuff.