suite: output live throughput data in exploded form
The live.csv file generated during a throughput test recorded the test identifications
data in an encoded string form (which is nice as it reduces the number of columns in
the file). This is ugly: While writing, while reading back and for external tools.
Thus output any datum in its own column.