suite: fix significant digits in througput test output
The start time of a Test was recorded on disk with 10 musec accuracy while deltas between
start times (column 'Elapsed') was output with 1 musec accuracy. This made offline
analysis (e.g. via 'cmd/ht stat') produce a different output as reading back the
live.csv file dropped musecs in the starttime. Fix this (at least in large parts)
by increasing the number of significant digits in the start time and reducing the
number of pronted digits for the calulated durations (here 10 musec is enough).