mock: enhance table driven variable setting
The old syntax allowed to map one variable to an other variable via a simple tyble.
The Hjson file you had to write did not read very well, the intention what gets mapped
to what based on which value was not very clear. And you could not express mappings like
  Set X to 'foo' if A=='bar' and B=='wuz'
The new schema allows exactly this and it reads more nicely and more naturaly.