suite: construct subsuite for mocks
Reporting the status if the mocks is a pain. The current solution was to log
stuff and to add autogenerated checks to the Test to report problems in the
mock. This will report what wnet wrong but is painful to debug as the very
useful information about whole request and whole response is not reported.

This CL generates a suite for the mock invocations (the actual and the expected)
and glues this as a metadata to the original Test.

Reporting is still missing.