suite: allow a mock to be disabled
The intended use for mocks inside a suite is to have a suite of tests
which work even if some thirdparty backend systems are unavailable.
The mocks provide these missing services. Now it might happen that what
actuall is missing depends on the envorinment the suite is executed:
E.g. on your local development laptop no backend services might be
available at all (so you need 3 mocks) while on some CI system there
might be one service available (requering 2 mocks) and on an integration
system all backend service are present and no mocks are needed.
Thus a mechanism to turn mocks on/off easily is needed. Variable
substitution is well suited for such envorinment specific configurations
so this CL adds a boolena flag to raw mocks to disable the mock.

It turned out to be a bit invasive. Error handling/reporting is still
bad.