ht: improve errors generated by JSON check
One of the main ideas behind ht is that failed test and checks should provide
good dignostic output which makes it easy to see what went wrong. Unfortunately
if the input to a JSON or JSONExpr check was malformed the error was correct
but did not contain line number or even just plain byte offset indication where
the error happened. This makes debugging unecesarily hard. So reuse old code
to add line numbers in failed JSON decoding.