suite: set filename for each test during iteration
The actual filename got lost during the laste refactoring: It now must be set
explicitely as a metadata to the test. The old code used the automatic
variable TEST_NAME which contains only the basename of the filename. The new
behavior is to set the full filename i.e. the full relative path: The old behavior
was okay as TEST_NAME and TEST_DIR had to be two different variables but now
we can do better. When complex suites are constructed where same-name tests
are included from different folders it is helpful to see the actual path.