ht: replace generic error types for checks with plaintext
The idea of having error type like NotFound or FoundForbidden to be returned by the
checks looked like a good idea but it tunred out that this type of error is just too
generic and might be missleading. E.g. a JSON check trying to find a JSON element
with path 'foo.2.bar' of type string and checking if this strings contains 'waz'
it might fail with NotFound and it is not clear whether the element was not found
or the substring 'waz' was not found inside the found string.
Some checks already did return special error messages but lots of checks embed a
Condition which return generic errors.

This CL gets rid of this generic errors and replaces them mostly with plain strings.
To get it started and buildable some tests for specific error types had to be
replaced with a test for 'any error at all' which makes the tests a bit less
significant. It might be sensible to replace most such generic 'any error' with
a specific error message.