suite: display error list as list in HTML report
The error from a check is already a ht.ErrorList but up to now the output was
the simple message generated by the Error() method i.e. a long string of individual
error messages seperated by ';'. Basically unreadable. Thist is HTML, an ul is
the perfect choice to list several error, e.g. problems found in a HTML file.