ht: collect check failures in a nice way into the test error
Up to now only the first error was propagated to the Test error field which
is okay but a bit useless: The first failing check is reported in a prominent
way and the others are burried in the individual checks.
This CL makes Test.Error a ErrorList in all cases and collects all check
errors augmented by the check name which is much more informative.