cmd/ht: run should produce a Main test
With the 'run' subcommand one can execute a single test and the ht tool will
autogenerate a suite around this test. Now such a run'ed test should be a
Main type test, not a Setup and especially not a Teardown test. Unfortunately
the code relies on the number of elements in Setup and Main to determine which
test is of which kind and if both are empty all tests are Teardwon tests.

Thus add a dummy element to Main, forcing the run'ed test to be a main test.