cmd/ht: restructure suite execution and reporting
Until this CL all suites were executed and only after executing them
all the results where written to disk (and displayed on stdout).
This makes it hard to observe the state of the suites tested.

With this CL this behavior is changed: Now each suite is executed
and safed to disk right after execution and the overall result HTML
file is regenerated to include this freshly finished suite. Like
this it is possible to monitor how a bunch of suites are executed
and what the state of the system under test is.

This might be even beneficial in regard to memory consumption as no
longer all responses from all tests from all suites have to be kept
in memory until the very end where they have been written to disk.

Some other aspects are changed too:
  - The timestamp of the output directory is now the time when the
    execution starts instead of when writing to disk starts.
    This seems more natural anyway.
  - Report printing of each suite happens right after the suite's
    execution. This makes monitoring the state on the terminal
    easier.
  - The text report is also saved to disk and not ony printed to
    stdout. Looking up the result of an archived suite on the
    command line was always awful as the only report was the
    HTML and the JUnit reports: Neither reads well.
  - All this file IO writing the reports might fail. Up to now
    these errors where either ignored or terminated to execution
    (which is bad but as all file IO happend after executing all
    suites at least the tests had been executed; one just did not
    see the result). With this CL all such errors are simply
    collected and reported in the very end. Such errors result now
    in an exit code of 7.