cmd/ht: fix execution of one suite twice
The report for a suite named <suitename> were written to a folder named
<suitename>. So if a suite is executed twice
    # ht exec some.suite some.suite
Then saving the second run will overwrite the results of the first one.
While this type of invoking ht might be very uncommon the problem occurs
also if two suites happen to have the same name.

This CL fixes this type of problem by prepending a serial number to
the generated folder name. It not only fixes the problem but has an extra
benefit: It becomes clear in which order the suites have been executed.