ht: make the Test an argument to checks Prepare method
The Execute method of a check always had to take a *Test as an argument for obvious
resons. The Prepare didn't. This is not a problem if Prepare is just used to
precompile a regexp. But if some checks are not suitable for some type of test
(e.g. Chache check on POST request) or if a checks wants to perform prepare work
based on details of the Test than this is not good. E.g. it is not noticed that
Cache is not aplicaple to a POST request Test until the check is executed. It could
be detected as bogus much earlier.

This breaks the API as the new Peparable interface has a different signature.
This change should have gone into the initial 'Prepare in its own interface'
change but I forgot it until cleaning up the TODO. Better late than never.