ht: remove two TODOs
The structure is generated via reflection and has to be writable. The easiest way
is creating an extractor with reflect.New, that works and produces a pointer
which is settable and can be populated. That's the reason all extractors are
pointer values after unmarshaling from Hjson.
Just remove the TODOs as all is okay.