fingerprint: use base-36 digits for string representation of color histograms
Base-32 (i.e. 0-9a-v) was used which seemded a an improvement from hex (base-16)
because it nicely doubles the encoded bits. But it is a stupid idea: The original
range in the histogram is [0,255] which is mapped to [0,31]. The mapping is
nonlinear to preserve a bit of information in the lower bits. The result is
a base32 number where 'v' is the larges digit which is unintuitiv.

This change uses base-36 (the largest package strconv can handle) which at
least maps the largest value to a 'z'. A followup CL should change this
to base64 (which still cuts of the 2 least significant digits).