fingerprint: consider colour perception when selecting palett color
Instead of the very crude euclidian distance in RGB use a simple approximation
(still in RGB) to capture color perception of humans.
Up to now all pixels were included in the color histogram, even if totaly
transparent. This is basically nonsense. The new code ignores pixels which
are too transparent.