cmd/ht: improve builtin help
Help for the types used in constructing Suites, Tests and Checks is good
and available via 'ht doc <type>' but cmd/ht does not really work with
the Go types but works on the Hjson files used to encode suites and tests.

This change reorganizes help and adds help on example suites, tests and mixins.