ht: use HTTP status codes in file:// pseudo-requests
If one of the file:// pseudo requests failed (e.g. because the path was not
accessible) it returned an error. This makes it complicated express some
intentions like "If /some/file does not exist then everything is fine
and the test should pass." as any file operation on a nonexisting path
would result in an error (and checks are not executed at all).

The HTTP status codes provide status for all these file operations
and their different outcomes and reporting the file operation results
as HTTP status codes makes expressing above cases very easy.

This change will allow to use the Error state for new classes of
problems: Once file:// pseudo request over ssh are implemented the
error cases can be used for errors like host not reachable or
bad credentials.

While at it fix documentation of sql:// pseudo requests.