ht: fix/improve bash pseudo request
The change to report more via HTTP status codes was a bit too eager:
The working directory for the execution is taken from the URL path
and the script execution fails badly if the working directory
does not exist (which is okay). Such an error was reported via a
500 Internal Server Error status code and an empty response body.
This is not helpful during debugging a broken test.

This change seperates stating the bash script from running it.
If it cannot be started at all this will be reported as an test
Error instead of a 500 Internal Server Error: This makes
debugging eaiser and is the more sensible choice as such a test
is at the border of beeing malformed (=Bogus) and we technically
did not make a request (=execute the script), we did not even start
the script.

This also fixes the tests I broke.